+++ DatadogRUM/Sources/Instrumentation/Resources/URLSessionRUMResourcesHandler.swift
@@ -154,12 +154,12 @@ extension DistributedTracing {
                 // To make sure the generated traces from RUM don’t affect APM Index Spans counts.
                 request.setValue("rum", forHTTPHeaderField: TracingHTTPHeaders.originField)
             case .b3:
-                writer = OTelHTTPHeadersWriter(
+                writer = B3HTTPHeadersWriter(
                     sampler: sampler,
                     injectEncoding: .single
                 )
             case .b3multi:
-                writer = OTelHTTPHeadersWriter(
+                writer = B3HTTPHeadersWriter(
                     sampler: sampler,
                     injectEncoding: .multiple
                 )